+++ tests/TestClasses/AggregateRoots/AccountAggregateRootThatAllowsConcurrency.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Spatie\EventSourcing\Tests\TestClasses\AggregateRoots;
-
-class AccountAggregateRootThatAllowsConcurrency extends AccountAggregateRoot
-{
-    protected static bool $allowConcurrency = true;
-}